Extract CloudConfig and mock for future PR
